Added processing of uni-direction nodes in distance matrix by float numbers

+++ src/test/java/com/hukuta94/pathnodebuilder/logic/parser/overwatch/OverwatchParserTest.java
@@ -118,7 +118,7 @@ public class OverwatchParserTest
                         {6, 5, 4}
                 };
 
-        int[][] distanceMatrix = new int[][]
+        float[][] distanceMatrix = new float[][]
                 {
                         {1, 2, 2, 2, 1, 1, 2},
                         {1, 2, 3, 2, 2, 3},
